+++ src/Eloquent/DefaultOrderScope.php
@@ -7,6 +7,9 @@ use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Scope;
 
+/**
+ * @deprecated in 5.0.0
+ */
 class DefaultOrderScope implements Scope
 {
     protected array $orders;
@@ -16,6 +19,8 @@ class DefaultOrderScope implements Scope
      */
     public function __construct(array $orders)
     {
+        trigger_error('The DefaultOrderScope() global scope should no longer be used, it has been deprecated since version 5.0.0; please implement a dedicated ordered() local scope instead.', E_USER_DEPRECATED);
+
         $this->orders = $orders;
     }