+++ src/Eloquent/Exceptions/JoinRelException.php
@@ -4,6 +4,4 @@ namespace Axn\Illuminate\Database\Eloquent\Exceptions;
 
 use Exception;
 
-class JoinRelException extends Exception
-{
-}
+class JoinRelException extends Exception {}