+++ src/Eloquent/Exceptions/WhereHasInException.php
@@ -4,6 +4,4 @@ namespace Axn\Illuminate\Database\Eloquent\Exceptions;
 
 use Exception;
 
-class WhereHasInException extends Exception
-{
-}
+class WhereHasInException extends Exception {}